staging random bag of poorly implemented shape space changed before trying to do same for transition shape change.

+++ Matt Seg GUI/@timelapseTraps/segmentACexperimental.m
@@ -376,7 +376,7 @@ for TP = Timepoints
                     end
                     
                     if TrapPresentBoolean
-                        ExpectedCellCentre = LocalExpectedCellCentre + [TrapLocations(trap).xcenter TrapLocations(trap).ycenter] - ([TrapWidth TrapHeight] + 1) ;
+                        ExpectedCellCentre = double(LocalExpectedCellCentre) + double([TrapLocations(trap).xcenter TrapLocations(trap).ycenter] - ([TrapWidth TrapHeight] + 1)) ;
                     else
                         ExpectedCellCentre = LocalExpectedCellCentre;
                     end
